refactor(app): consolidate message action lifecycle
